Hacked a SMART app launch POC
This integrates with https://github.com/brianpos/smart-on-fhir.  It is a Windows app acting as a EHR to launch SMART Apps.  The only thing I had to do to get it to run on my Windows 11 box was to add the following binding redirect:

  <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
    <dependentAssembly>
      <assemblyIdentity name="System.Net.Http" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
      <bindingRedirect oldVersion="0.0.0.0-4.8.0.0" newVersion="4.8.0.0" />
    </dependentAssembly>
  </assemblyBinding>
</assemblyBinding>

Then in UdapEd I created a launchBP.razor page as the main launch.  In the "SmartSession from.well - known / smart - configuration" section click the CONTINUE TO AUTHORIZE button to get an access code.  Then click the  REQUEST TOKEN button to get an access token.  The token is placed in your user session so you can use it in the Patient Search Page.  But because the PatientSearch searches from the backend rather than WASM you can't use the https://fhir.test.localhost URL.  But you can use the one auto generated at startup based on port.  You will it in the bottom status area when the EHRApp starts up.