Build LauncContext from hl7_identitifer
The problem with this is the hl7_identifier is a FHIR Identifier in my GCP FHIR server.  When using Patient prefixed scopes GCP will require a X-Authorization-Patient header to set the patient context.  The value must be the Patient ID.  So I am not sure I will use this technique.  I would have to then find the Patient on a back channel assuming I had a patient with that hl7_identifier.  I think the UdapEd tool would need to set up a process where we look at patients in  our own FHIR server then from one of those patients match on a patient from another FHIR Server and if the match is good enough then set the Patient Context through some means.

Just want to keep this code around for a while.