Fix an issue with Maui
It does not create a new IDiscoveryService for every call to GetUdapMetadataVerificationModel so the events registered with _udapClient would keep stacking up and all items in the collection would recieve and event notification from any one of them.  So it would appears all UDAP endpoints failed to validate.  Only in Mauil and not in Blazor web.