mTLS Client on Windows
Experimenting more with Windows and mTLS client.  It just isn't possible to completely override certificate validation on Windows without putting the CA in the Windows Certificate Store.  Interesting that it works find on Linux.  Might be interesting to pull the the .NET core source and troubleshoot some more.  Could be a MS PR.  But I suspect I will spend two days on it.  Thinking....