Update obsolete code and enabled static content fingerprinting
The new X509Certificate2 technique is obsolete along with some usages of new UdapCertificate2Collection.  Updated all of these.
Also added change app.UseStaticFiles() to app.MapStaticAssets() which is middleware to fingerprint static content so the client will know when to the static content is new it's cache needs to update.