WinUIEx working
I tried this out on a Windows 10 machine rather than my typical Windows 11 machine.  Not sure why but on the Windows 10 machine, running that app indicating :
Microsoft.AspNetCore.Components.Web.ErrorBoundary: Error: System.InvalidOperationException: The WebAuthenticator requires a packaged app with an AppxManifest

This was not happening on the Windows 11 box.  In addition I  needed to set a WindowsPackageType to MSIX and then in the launchsettings I needed to set the command to MsixPackage.

I am not 100% in the know yet on the packaging process and deploying the Maui app.  Going to try this code out on the Windows 11 box and see what is up.

Side note:  I am also struggling with running on MacCatalyst at the moment because I can't sign my code.  That is most likely due to not being able to connect my IOS device to a corporate laptop.  Working on that.