IOS and MacCatalyst wip
Drop the TabbedPage for now.  It just make it more difficult to control the padding of the app so that it does not land in the safe area of the iPhone.  I did bring in the Device Orientation work which would make this easier to fixup the TabbedPage but for now I am just don't want to deal with the complexity this adds as I try to get this to work on IOS and MacCatalyst.